requirements.txt
asgiref==3.7.2
Django==4.2.7
sqlparse==0.4.4
typing_extensions==4.8.0
djangorestframework==3.14.0
